implement CRUD operations to vehicle make

+++ Service-Throttle-Backend/service-throttle-backend/src/main/java/com/servicethrottle/servicethrottlebackend/models/MobileServiceVehicle.java
@@ -17,7 +17,7 @@ import static javax.persistence.FetchType.LAZY;
 @NoArgsConstructor
 public class MobileServiceVehicle extends Vehicle {
 
-    private long capacity;
+    private double capacity;
 
     @OneToOne(fetch = LAZY)
     private MobileMechanic mobileMechanic;
